chore: Standardized the file path for Registry
