--- types/glint.d.ts
@@ -1,2 +1,0 @@
-import '@glint/environment-ember-loose';
-import 'ember-intl/glint';
